% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booticer.R
\docType{package}
\name{booticer}
\alias{booticer}
\alias{booticer-package}
\title{booticer: Get bootstrapped ICER}
\description{
`booticer` is a R package created to easily produce cost-effectiveness planes and
acceptability curves
}
