% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{get_differences}
\alias{get_differences}
\title{Get cost and effects difference in each strategy}
\usage{
get_differences(x, reference)
}
\arguments{
\item{x}{A list created with get_means}

\item{reference}{A atomic character vector naming the reference strategy}
}
\value{
A data.frame with one line for each resample
}
\description{
Get cost and effects difference in each strategy
}
\examples{

}
