% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{get_means}
\alias{get_means}
\title{Get the mean cost and effect in each strategy}
\usage{
get_means(res)
}
\arguments{
\item{res}{A list created with get_resamples}
}
\value{
A list with a length equal to the number of strategies containing a
data.frame with one line for each resample
}
\description{
Get the mean cost and effect in each strategy
}
\examples{

}
