% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{get_quadrant_prop}
\alias{get_quadrant_prop}
\title{Get proportion of bootstrapped ICERs in each quadrant}
\usage{
get_quadrant_prop(x, strategy)
}
\arguments{
\item{x}{an object created with get_differences function}

\item{strategy}{the strategy that you want to analyse}
}
\description{
Get proportion of bootstrapped ICERs in each quadrant
}
