% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{get_resamples}
\alias{get_resamples}
\title{Create non-parametric resamples with replacement of a data.frame}
\usage{
get_resamples(tab, n = 1000, seed = NULL)
}
\arguments{
\item{tab}{A data frame with colnames containing "cost", "effect", and "strategy"}

\item{n}{Number of resamples}

\item{seed}{Parameter to reproduce the results}
}
\value{
A list with a length equal to the number of strategies containing a list of n
data frames
}
\description{
Create non-parametric resamples with replacement of a data.frame
}
\examples{

}
